新增 E 模組賣家中心功能

+++ backend/src/main/java/com/dinogo/coupon/dto/CouponCreateRequest.java
@@ -16,7 +16,7 @@ public record CouponCreateRequest(
         @NotNull LocalDateTime endAt,
         Integer limitCount,
         @NotBlank String scopeType,
-        Long categoryId,
-        Long productId
+        Integer categoryId,
+        Integer productId
 ) {
 }
